fixup! Change light type

+++ color.cpp
@@ -3,5 +3,7 @@
 
 std::ostream& operator<<(std::ostream &out , const Color &color)
 {
-    return out << "(" << color.RED << ", " << color.GREEN << ", " << color.BLUE << ")";
+    return out << "("  << (short) color.RED
+               << ", " << (short) color.GREEN
+               << ", " << (short) color.BLUE << ")";
 }
